backend/requirements.txt
fastapi
uvicorn
pycparser
pydantic
python-multipart
django
djangorestframework
django-cors-headers